--- bin/employeedb/view/EmployeeDB.fxml
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.*?>
-<?import javafx.scene.control.*?>
-<?import java.lang.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.layout.AnchorPane?>
-
-<AnchorPane prefHeight="400.0" prefWidth="300.0" xmlns="http://javafx.com/javafx/8.0.40" xmlns:fx="http://javafx.com/fxml/1" fx:controller="employeedb.view.EmployeeDBController">
-   <children>
-      <TableView fx:id="employeeTable" layoutX="28.0" layoutY="24.0" prefHeight="200.0" prefWidth="200.0" AnchorPane.bottomAnchor="100.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
-        <columns>
-          <TableColumn fx:id="firstNameColumn" prefWidth="75.0" text="First Name" />
-          <TableColumn fx:id="lastNameColumn" prefWidth="75.0" text="Last Name" />
-        </columns>
-         <columnResizePolicy>
-            <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
-         </columnResizePolicy>
-      </TableView>
-      <GridPane layoutX="6.0" layoutY="300.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="320.0">
-        <columnConstraints>
-          <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-          <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-            <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-            <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-            <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-            <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-        </columnConstraints>
-        <rowConstraints>
-          <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-        </rowConstraints>
-         <children>
-            <Label text="First">
-               <GridPane.margin>
-                  <Insets left="5.0" />
-               </GridPane.margin>
-            </Label>
-            <Label fx:id="firstNameLabel" text="Label" GridPane.columnIndex="1" />
-            <Label text="Last" GridPane.columnIndex="2" />
-            <Label fx:id="lastNameLabel" text="Label" GridPane.columnIndex="3" />
-            <Label text="Salary" GridPane.columnIndex="4" />
-            <Label fx:id="salaryLabel" text="Label" GridPane.columnIndex="5" />
-         </children>
-      </GridPane>
-      <ButtonBar layoutX="86.0" layoutY="346.0" prefHeight="40.0" prefWidth="200.0" AnchorPane.bottomAnchor="10.0" AnchorPane.rightAnchor="10.0">
-        <buttons>
-          <Button mnemonicParsing="false" onAction="#handleNewEmployee" text="New" />
-            <Button mnemonicParsing="false" onAction="#handleEditEmployee" text="Edit" />
-            <Button mnemonicParsing="false" onAction="#handleDeleteEmployee" text="Delete" />
-        </buttons>
-      </ButtonBar>
-   </children>
-</AnchorPane>
