+++ bin/tetris/view/LeaderBoard.fxml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.text.*?>
+<?import javafx.scene.control.*?>
+<?import java.lang.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.scene.layout.AnchorPane?>
+
+
+<AnchorPane prefHeight="300.0" prefWidth="400.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.40">
+   <children>
+      <TableView prefHeight="250.0" prefWidth="400.0" AnchorPane.topAnchor="50.0">
+        <columns>
+          <TableColumn prefWidth="133.0" text="Name" />
+          <TableColumn prefWidth="133.0" text="Level" />
+            <TableColumn prefWidth="133.0" text="Score" />
+        </columns>
+      </TableView>
+      <Label alignment="CENTER" layoutX="133.0" layoutY="14.0" text="LEADER BOARD">
+         <font>
+            <Font size="18.0" />
+         </font>
+      </Label>
+   </children>
+</AnchorPane>
