+++ src/tetris/view/MainMenuController.java
@@ -1,0 +1,31 @@
+package tetris.view;
+
+import javafx.fxml.FXML;
+import javafx.scene.control.Label;
+
+public class MainMenuController 
+{
+	@FXML
+	private Label start;
+	
+	@FXML
+	private Label resume;
+	
+	@FXML
+	private Label leader;
+	
+	@FXML
+	private Label exit;
+	
+	@FXML
+	private void initialize()
+	{
+		
+	}
+	
+	@FXML
+	private void handleStartGame()
+	{
+		
+	}
+}
